Update field handlers
The previous version was spliting the fields on the "." and recreating
a dictionary. For instance:

    ['user.username', 'user.stats.monthly']

Was generating:

    {'user': {
        'username': None,
        'stats': {
            'monthly': None
        }}

The issue: when you fetch a model, you pass an array: ['user.username', ...],
so if you need to fetch again on an extension, you should also pass the same
signature.

Now, the flow is a lot easier to understand:

     [user.username, user.stats.monthly]

     calls: user and pass the above, find the username and fetch the extension
            stats and pass to stats: ['stats.monthly']